feat(dlna): rewrite ContentDirectory on Subsonic HTTP, no DB (closes #219)
Phase 5 of #212. The DLNA ContentDirectory service no longer touches
`better-sqlite3` directly — every browse + search call routes through a
`SubsonicCaller` against the hub's own `/rest/*` endpoints over in-process
`app.inject()`. Promotes the #213 spike to production and deletes the
parallel module + spike test.

- Replace `DlnaObjectService(db)` with `DlnaObjectService(SubsonicCaller)`;
  browse + search are now async.
- Wire the production caller in `server.ts` using owner u+p over loopback
  (`app.inject`). HTTP-shaped interface keeps the future split-deploy
  swap to real loopback fetch trivial.
- Drop the "All Tracks" container (`0/music/tracks`) from the hierarchy.
  Subsonic has no global track endpoint; the only synthesis is N+1 over
  getAlbumList2+getAlbum which is unacceptable at scale. WMP/Kodi/VLC/
  BubbleUPnP all enter Artists or Albums by default — cosmetic loss only.
  A future OpenSubsonic `getSongs` extension (#214) can restore it cheap.
- `childCount=-1` on the synthetic Artists/Albums containers and on the
  global Albums browse so we don't pay extra round-trips just to populate
  exact counts (clients tolerate -1 and walk the pager).
- Promote `hub/test/dlna-subsonic-spike.test.ts` into the production
  `hub/test/dlna-objects.test.ts`; delete the spike file + module.
- Update `docs/dlna.md` (hierarchy, "All Tracks" rationale, op→Subsonic
  mapping, new test wiring) and `docs/system-architecture.md`.

Verify: `pnpm verify` green (577 hub tests, 92 frontend, 49 files);
`pnpm --filter hub test:integration -- dlna-http.integration` green
(9 tests). Pre-existing SSDP integration failures are environmental
(UDP 1900 multicast) and unrelated to this change.
