+++ hub/src/services/dlna-objects.ts
@@ -1,28 +1,41 @@
 /**
- * DLNA ContentDirectory object hierarchy backed by the unified library.
+ * DLNA ContentDirectory object hierarchy backed by the hub's own Subsonic
+ * API (#219, Phase 5 of #212).
+ *
+ * No direct database access. Every browse / search call resolves through a
+ * `SubsonicCaller` — typically an in-process `app.inject()` wrapper, but
+ * the interface is HTTP-shaped so a deployment split (Player as a separate
+ * process talking to the Hub over loopback fetch) is a swap of the caller,
+ * not a rewrite of this module.
  *
  * Object IDs are deterministic strings so DLNA clients (notably Windows
  * Media Player) can cache them across hub restarts without breaking. They
- * survive resync because they're keyed on `unified_*.id` UUIDs, which are
- * stable for the lifetime of an entity.
+ * survive resync because they're keyed on `unified_*.id` UUIDs (the Subsonic
+ * `ar`/`al`/`t` prefix is stripped on the way in and re-added on outbound
+ * calls).
  *
  * Tree:
  *
- *   "0"                        root container
- *     "0/music"                Music
- *       "0/music/artists"      All Artists
- *         "0/music/artist/<unified_artist_id>"   one artist
+ *   "0"                                  root
+ *     "0/music"                          Music
+ *       "0/music/artists"                All Artists
+ *         "0/music/artist/<unified_artist_id>"        one artist
  *           — contains release groups for that artist
- *       "0/music/albums"       All Albums (release groups)
- *         "0/music/album/<unified_release_group_id>"   one album
+ *       "0/music/albums"                 All Albums (release groups)
+ *         "0/music/album/<unified_release_group_id>"  one album
  *           — contains tracks across all releases in the group
- *       "0/music/tracks"       All Tracks
  *
- * We collapse `unified_release_groups` to "albums" for the v1 hierarchy;
- * release-level (edition) browsing is out of scope and rarely useful for
- * DLNA clients.
+ * "All Tracks" (a global flat track enumeration) is intentionally NOT
+ * exposed: Subsonic has no `getTracks` endpoint, and the only way to
+ * synthesize it on top of `getAlbumList2 + getAlbum` is an N+1 fan-out.
+ * Per spike #213 recommendation we drop the container. WMP / Kodi / VLC /
+ * BubbleUPnP all default to entering Artists or Albums anyway. A future
+ * OpenSubsonic `getSongs` extension (#214 survey) would let us add it back
+ * cheaply; until then this is a documented gap.
+ *
+ * Release-level (edition) browsing is also not exposed — `unified_release_groups`
+ * is the natural "album" object for DLNA clients.
  */
-import type Database from "better-sqlite3";
 import { buildAudioItem, buildContainer, wrapDidl } from "./didl.js";
 import { xmlEscape } from "./soap.js";
 import { buildStreamUrl } from "./cast-tokens.js";
@@ -31,14 +44,12 @@ export const ROOT_ID = "0";
 export const MUSIC_ID = "0/music";
 export const ARTISTS_ID = "0/music/artists";
 export const ALBUMS_ID = "0/music/albums";
-export const TRACKS_ID = "0/music/tracks";
 
 export type ObjectKind =
   | "root"
   | "music"
   | "artists"
   | "albums"
-  | "tracks"
   | "artist"
   | "album";
 
@@ -53,7 +64,6 @@ export function parseObjectId(s: string): ParsedObjectId | null {
   if (s === MUSIC_ID) return { kind: "music" };
   if (s === ARTISTS_ID) return { kind: "artists" };
   if (s === ALBUMS_ID) return { kind: "albums" };
-  if (s === TRACKS_ID) return { kind: "tracks" };
   const mArtist = s.match(/^0\/music\/artist\/([^/]+)$/);
   if (mArtist) return { kind: "artist", id: mArtist[1] };
   const mAlbum = s.match(/^0\/music\/album\/([^/]+)$/);
@@ -79,7 +89,7 @@ export interface BrowseOptions {
    */
   baseUrl: string;
   /**
-   * Cast-token signer (#218). `res@uri` is now a self-contained
+   * Cast-token signer (#218). `res@uri` is a self-contained
    * `${baseUrl}/rest/stream.view?id=…&castToken=…&dlna=1` URL so the DLNA
    * renderer pulls bytes directly from the Hub Subsonic endpoint. No
    * Player relay, no per-renderer credentials.
@@ -104,66 +114,105 @@ export interface BrowseResult {
   result: string;
   /** Number of objects returned in this response. */
   numberReturned: number;
-  /** Total number of child objects (for pagination). */
+  /**
+   * Total number of child objects (for pagination). `-1` means "unknown" —
+   * Subsonic `getAlbumList2` does not return a total, and UPnP clients
+   * tolerate -1 by walking the pager until they get a short page.
+   */
   totalMatches: number;
 }
 
-interface ArtistRow {
-  id: string;
-  name: string;
-  image_url: string | null;
-  album_count: number;
+/**
+ * Minimal Subsonic JSON client surface the DLNA service needs. Production
+ * uses an `app.inject()`-backed implementation (in-process loopback, no TCP).
+ * Tests substitute fixture or mock callers.
+ */
+export interface SubsonicCaller {
+  /** Returns the decoded JSON `subsonic-response` body. Throws on non-200. */
+  call(endpoint: string, params: Record<string, string>): Promise<SubsonicResponse>;
 }
 
-interface AlbumRow {
-  id: string;
-  name: string;
-  artist_id: string;
-  artist_name: string;
-  image_url: string | null;
-  track_count: number;
+export interface SubsonicResponse {
+  "subsonic-response": Record<string, unknown> & {
+    status: "ok" | "failed";
+    error?: { code: number; message: string };
+  };
 }
 
-interface TrackRow {
+interface SubsonicArtistIndex {
+  ignoredArticles: string;
+  index: Array<{
+    name: string;
+    artist: Array<{
+      id: string;
+      name: string;
+      albumCount?: number;
+      coverArt?: string;
+    }>;
+  }>;
+}
+
+interface SubsonicArtist {
+  id: string;
+  name: string;
+  albumCount: number;
+  coverArt?: string;
+  album: SubsonicAlbumSummary[];
+}
+
+interface SubsonicAlbumSummary {
+  id: string;
+  name: string;
+  artist: string;
+  artistId: string;
+  coverArt?: string;
+  songCount: number;
+  year?: number;
+}
+
+interface SubsonicAlbum extends SubsonicAlbumSummary {
+  song: SubsonicSong[];
+}
+
+interface SubsonicSong {
   id: string;
   title: string;
-  artist_name: string;
-  album_name: string;
-  album_art: string | null;
-  duration_ms: number | null;
-  format: string | null;
-  bitrate: number | null;
+  album: string;
+  artist: string;
+  albumId?: string;
+  artistId?: string;
+  coverArt?: string;
+  duration?: number;
+  bitRate?: number;
+  contentType?: string;
+  suffix?: string;
+  track?: number;
+  discNumber?: number;
 }
 
-/**
- * Maps the unified library's source format hint to a DLNA-suitable MIME
- * type. Returns a best-effort `audio/mpeg` if unknown — Windows Media
- * Player legacy in particular wants something here.
- */
-function mimeForFormat(format: string | null | undefined): string {
-  switch ((format || "").toLowerCase()) {
-    case "mp3":
-      return "audio/mpeg";
-    case "flac":
-      return "audio/flac";
-    case "ogg":
-    case "opus":
-      return "audio/ogg";
-    case "aac":
-    case "m4a":
-      return "audio/mp4";
-    case "wav":
-      return "audio/wav";
-    default:
-      return "audio/mpeg";
-  }
+interface SubsonicSearch3 {
+  artist?: Array<{ id: string; name: string; albumCount?: number; coverArt?: string }>;
+  album?: SubsonicAlbumSummary[];
+  song?: SubsonicSong[];
 }
 
-function streamUri(opts: BrowseOptions, trackId: string): string {
+/** Strip the `ar`, `al`, or `t` prefix from a Subsonic-encoded id. */
+function stripPrefix(id: string, prefix: string): string {
+  return id.startsWith(prefix) ? id.slice(prefix.length) : id;
+}
+
+function mimeForContentType(ct: string | undefined): string {
+  // Subsonic returns per-track Content-Type for the source file (mp3 →
+  // audio/mpeg, flac → audio/flac, …). Fall back to audio/mpeg when the
+  // upstream omits it — Windows Media Player legacy needs something here.
+  return ct && ct.length > 0 ? ct : "audio/mpeg";
+}
+
+function streamUri(opts: BrowseOptions, unifiedTrackId: string): string {
   return buildStreamUrl({
     lanUrl: opts.baseUrl,
     castSecret: opts.castSecret,
-    unifiedTrackId: trackId,
+    unifiedTrackId,
     username: opts.username,
     ttlSec: opts.ttlSec,
     client: "poutine-dlna",
@@ -171,20 +220,25 @@ function streamUri(opts: BrowseOptions, trackId: string): string {
   });
 }
 
-function albumArtUriFor(baseUrl: string, encodedCoverArtId: string | null): string | null {
-  if (!encodedCoverArtId) return null;
-  return `${baseUrl}/rest/getCoverArt?id=${encodeURIComponent(encodedCoverArtId)}`;
+function coverArtUri(baseUrl: string, coverArtId: string | undefined | null): string | null {
+  if (!coverArtId) return null;
+  // Cover-art ids returned by Subsonic responses are already in the
+  // `{instanceId}:{coverArtId}` shape expected by `/rest/getCoverArt`.
+  return `${baseUrl}/rest/getCoverArt?id=${encodeURIComponent(coverArtId)}`;
 }
 
+const DLNA_PROTOCOL_INFO_EXTRAS =
+  "DLNA.ORG_OP=01;DLNA.ORG_FLAGS=01700000000000000000000000000000";
+
 export class DlnaObjectService {
-  constructor(private readonly db: Database.Database) {}
+  constructor(private readonly caller: SubsonicCaller) {}
 
   /** ContentDirectory `Browse` (BrowseDirectChildren or BrowseMetadata). */
-  browse(
+  async browse(
     objectId: string,
     browseFlag: "BrowseMetadata" | "BrowseDirectChildren",
     opts: BrowseOptions,
-  ): BrowseResult {
+  ): Promise<BrowseResult> {
     const parsed = parseObjectId(objectId);
     if (!parsed) {
       return { result: wrapDidl(""), numberReturned: 0, totalMatches: 0 };
@@ -195,85 +249,135 @@ export class DlnaObjectService {
     return this.browseChildren(parsed, opts);
   }
 
-  private browseMetadata(parsed: ParsedObjectId, opts: BrowseOptions): BrowseResult {
-    const xml = (() => {
-      switch (parsed.kind) {
-        case "root":
-          return buildContainer({
-            objectId: ROOT_ID,
-            parentId: "-1",
-            title: "Poutine",
-            upnpClass: "object.container.storageFolder",
-            childCount: 1,
-          });
-        case "music":
-          return buildContainer({
-            objectId: MUSIC_ID,
-            parentId: ROOT_ID,
-            title: "Music",
-            upnpClass: "object.container.storageFolder",
-            childCount: 3,
-          });
-        case "artists":
-          return buildContainer({
-            objectId: ARTISTS_ID,
-            parentId: MUSIC_ID,
-            title: "Artists",
-            upnpClass: "object.container.person.musicArtist",
-            childCount: this.countArtists(),
-          });
-        case "albums":
-          return buildContainer({
-            objectId: ALBUMS_ID,
-            parentId: MUSIC_ID,
-            title: "Albums",
-            upnpClass: "object.container.album.musicAlbum",
-            childCount: this.countAlbums(),
-          });
-        case "tracks":
-          return buildContainer({
-            objectId: TRACKS_ID,
-            parentId: MUSIC_ID,
-            title: "All Tracks",
-            upnpClass: "object.container.storageFolder",
-            childCount: this.countTracks(),
-          });
-        case "artist": {
-          const row = this.getArtist(parsed.id!);
-          if (!row) return "";
-          return buildContainer({
-            objectId: artistObjectId(row.id),
-            parentId: ARTISTS_ID,
-            title: row.name,
-            upnpClass: "object.container.person.musicArtist",
-            childCount: row.album_count,
-            albumArtUri: row.image_url,
-            artist: row.name,
-          });
-        }
-        case "album": {
-          const row = this.getAlbum(parsed.id!);
-          if (!row) return "";
-          return buildContainer({
-            objectId: albumObjectId(row.id),
-            parentId: ALBUMS_ID,
-            title: row.name,
-            upnpClass: "object.container.album.musicAlbum",
-            childCount: row.track_count,
-            albumArtUri: albumArtUriFor(opts.baseUrl, row.image_url),
-            artist: row.artist_name,
-          });
-        }
-      }
-    })();
+  /** ContentDirectory `Search`. Maps to Subsonic search3. */
+  async search(query: string, opts: BrowseOptions): Promise<BrowseResult> {
+    const size = opts.requestedCount > 0 ? Math.min(opts.requestedCount, 100) : 50;
+    const body = await this.caller.call("/rest/search3", {
+      query,
+      artistCount: String(size),
+      albumCount: String(size),
+      songCount: String(size),
+    });
+    const res = (body["subsonic-response"].searchResult3 as SubsonicSearch3 | undefined) ?? {};
+    const items: string[] = [];
+    for (const a of res.artist ?? []) {
+      items.push(
+        buildContainer({
+          objectId: artistObjectId(stripPrefix(a.id, "ar")),
+          parentId: ARTISTS_ID,
+          title: a.name,
+          upnpClass: "object.container.person.musicArtist",
+          childCount: a.albumCount ?? -1,
+          albumArtUri: coverArtUri(opts.baseUrl, a.coverArt),
+          artist: a.name,
+        }),
+      );
+    }
+    for (const al of res.album ?? []) {
+      items.push(
+        buildContainer({
+          objectId: albumObjectId(stripPrefix(al.id, "al")),
+          parentId: ALBUMS_ID,
+          title: al.name,
+          upnpClass: "object.container.album.musicAlbum",
+          childCount: al.songCount,
+          albumArtUri: coverArtUri(opts.baseUrl, al.coverArt),
+          artist: al.artist,
+        }),
+      );
+    }
+    for (const s of res.song ?? []) {
+      items.push(this.songItem(s, /*parentId*/ ALBUMS_ID, opts));
+    }
     return {
-      result: wrapDidl(xml),
-      numberReturned: xml ? 1 : 0,
-      totalMatches: xml ? 1 : 0,
+      result: wrapDidl(items.join("")),
+      numberReturned: items.length,
+      totalMatches: items.length,
     };
   }
 
-  private browseChildren(parsed: ParsedObjectId, opts: BrowseOptions): BrowseResult {
+  // ── BrowseMetadata ─────────────────────────────────────────────────────
+
+  private async browseMetadata(
+    parsed: ParsedObjectId,
+    opts: BrowseOptions,
+  ): Promise<BrowseResult> {
+    let xml = "";
+    switch (parsed.kind) {
+      case "root":
+        xml = buildContainer({
+          objectId: ROOT_ID,
+          parentId: "-1",
+          title: "Poutine",
+          upnpClass: "object.container.storageFolder",
+          childCount: 1,
+        });
+        break;
+      case "music":
+        xml = buildContainer({
+          objectId: MUSIC_ID,
+          parentId: ROOT_ID,
+          title: "Music",
+          upnpClass: "object.container.storageFolder",
+          childCount: 2,
+        });
+        break;
+      case "artists":
+        xml = buildContainer({
+          objectId: ARTISTS_ID,
+          parentId: MUSIC_ID,
+          title: "Artists",
+          upnpClass: "object.container.person.musicArtist",
+          childCount: -1,
+        });
+        break;
+      case "albums":
+        xml = buildContainer({
+          objectId: ALBUMS_ID,
+          parentId: MUSIC_ID,
+          title: "Albums",
+          upnpClass: "object.container.album.musicAlbum",
+          childCount: -1,
+        });
+        break;
+      case "artist": {
+        const artist = await this.fetchArtist(parsed.id!);
+        if (!artist) break;
+        xml = buildContainer({
+          objectId: artistObjectId(parsed.id!),
+          parentId: ARTISTS_ID,
+          title: artist.name,
+          upnpClass: "object.container.person.musicArtist",
+          childCount: artist.albumCount,
+          albumArtUri: coverArtUri(opts.baseUrl, artist.coverArt),
+          artist: artist.name,
+        });
+        break;
+      }
+      case "album": {
+        const album = await this.fetchAlbum(parsed.id!);
+        if (!album) break;
+        xml = buildContainer({
+          objectId: albumObjectId(parsed.id!),
+          parentId: ALBUMS_ID,
+          title: album.name,
+          upnpClass: "object.container.album.musicAlbum",
+          childCount: album.song.length,
+          albumArtUri: coverArtUri(opts.baseUrl, album.coverArt),
+          artist: album.artist,
+        });
+        break;
+      }
+    }
+    return { result: wrapDidl(xml), numberReturned: xml ? 1 : 0, totalMatches: xml ? 1 : 0 };
+  }
+
+  // ── BrowseDirectChildren ───────────────────────────────────────────────
+
+  private async browseChildren(
+    parsed: ParsedObjectId,
+    opts: BrowseOptions,
+  ): Promise<BrowseResult> {
     switch (parsed.kind) {
       case "root":
         return this.staticChildren(
@@ -283,12 +387,15 @@ export class DlnaObjectService {
               parentId: ROOT_ID,
               title: "Music",
               upnpClass: "object.container.storageFolder",
-              childCount: 3,
+              childCount: 2,
             }),
           ],
           opts,
         );
       case "music":
+        // Static "Music" subtree. childCount=-1 (unknown) lets us skip the
+        // round-trips Subsonic would need to compute exact counts; every UPnP
+        // client we care about tolerates -1 and just walks in.
         return this.staticChildren(
           [
             buildContainer({
@@ -296,21 +403,14 @@ export class DlnaObjectService {
               parentId: MUSIC_ID,
               title: "Artists",
               upnpClass: "object.container.storageFolder",
-              childCount: this.countArtists(),
+              childCount: -1,
             }),
             buildContainer({
               objectId: ALBUMS_ID,
               parentId: MUSIC_ID,
               title: "Albums",
               upnpClass: "object.container.storageFolder",
-              childCount: this.countAlbums(),
-            }),
-            buildContainer({
-              objectId: TRACKS_ID,
-              parentId: MUSIC_ID,
-              title: "All Tracks",
-              upnpClass: "object.container.storageFolder",
-              childCount: this.countTracks(),
+              childCount: -1,
             }),
           ],
           opts,
@@ -319,13 +419,142 @@ export class DlnaObjectService {
         return this.listArtists(opts);
       case "albums":
         return this.listAlbums(null, opts);
-      case "tracks":
-        return this.listTracks(null, null, opts);
       case "artist":
         return this.listAlbums(parsed.id!, opts);
       case "album":
-        return this.listTracks(null, parsed.id!, opts);
+        return this.listTracks(parsed.id!, opts);
     }
+  }
+
+  // ── List builders ──────────────────────────────────────────────────────
+
+  private async listArtists(opts: BrowseOptions): Promise<BrowseResult> {
+    const body = await this.caller.call("/rest/getArtists", {});
+    const idx = (body["subsonic-response"].artists as SubsonicArtistIndex | undefined) ?? {
+      ignoredArticles: "",
+      index: [],
+    };
+    const all = idx.index.flatMap((g) => g.artist);
+    const total = all.length;
+    const sliced = all.slice(opts.startIndex, opts.startIndex + (opts.requestedCount || all.length));
+    const xml = sliced
+      .map((a) =>
+        buildContainer({
+          objectId: artistObjectId(stripPrefix(a.id, "ar")),
+          parentId: ARTISTS_ID,
+          title: a.name,
+          upnpClass: "object.container.person.musicArtist",
+          childCount: a.albumCount ?? -1,
+          albumArtUri: coverArtUri(opts.baseUrl, a.coverArt),
+          artist: a.name,
+        }),
+      )
+      .join("");
+    return { result: wrapDidl(xml), numberReturned: sliced.length, totalMatches: total };
+  }
+
+  private async listAlbums(
+    rawArtistId: string | null,
+    opts: BrowseOptions,
+  ): Promise<BrowseResult> {
+    if (rawArtistId) {
+      // Artist subtree — single call returns the album list ordered
+      // year DESC, name (Subsonic getArtist).
+      const body = await this.caller.call("/rest/getArtist", { id: `ar${rawArtistId}` });
+      const artist = body["subsonic-response"].artist as SubsonicArtist | undefined;
+      if (!artist) {
+        return { result: wrapDidl(""), numberReturned: 0, totalMatches: 0 };
+      }
+      const total = artist.album?.length ?? 0;
+      const sliced = (artist.album ?? []).slice(
+        opts.startIndex,
+        opts.startIndex + (opts.requestedCount || total),
+      );
+      const xml = sliced
+        .map((a) =>
+          buildContainer({
+            objectId: albumObjectId(stripPrefix(a.id, "al")),
+            parentId: artistObjectId(rawArtistId),
+            title: a.name,
+            upnpClass: "object.container.album.musicAlbum",
+            childCount: a.songCount,
+            albumArtUri: coverArtUri(opts.baseUrl, a.coverArt),
+            artist: a.artist,
+          }),
+        )
+        .join("");
+      return { result: wrapDidl(xml), numberReturned: sliced.length, totalMatches: total };
+    }
+
+    // Global "Albums" — paginate via getAlbumList2. Subsonic doesn't return
+    // a total in this response, so totalMatches stays -1 and the client
+    // walks the pager until it gets a short page.
+    const size = opts.requestedCount > 0 ? Math.min(opts.requestedCount, 500) : 500;
+    const body = await this.caller.call("/rest/getAlbumList2", {
+      type: "alphabeticalByName",
+      size: String(size),
+      offset: String(opts.startIndex),
+    });
+    const list =
+      ((body["subsonic-response"].albumList2 as { album?: SubsonicAlbumSummary[] } | undefined)
+        ?.album) ?? [];
+    const xml = list
+      .map((a) =>
+        buildContainer({
+          objectId: albumObjectId(stripPrefix(a.id, "al")),
+          parentId: ALBUMS_ID,
+          title: a.name,
+          upnpClass: "object.container.album.musicAlbum",
+          childCount: a.songCount,
+          albumArtUri: coverArtUri(opts.baseUrl, a.coverArt),
+          artist: a.artist,
+        }),
+      )
+      .join("");
+    return { result: wrapDidl(xml), numberReturned: list.length, totalMatches: -1 };
+  }
+
+  private async listTracks(rawAlbumId: string, opts: BrowseOptions): Promise<BrowseResult> {
+    const album = await this.fetchAlbum(rawAlbumId);
+    if (!album) return { result: wrapDidl(""), numberReturned: 0, totalMatches: 0 };
+    const total = album.song.length;
+    const sliced = album.song.slice(
+      opts.startIndex,
+      opts.startIndex + (opts.requestedCount || total),
+    );
+    const parent = albumObjectId(rawAlbumId);
+    const xml = sliced.map((s) => this.songItem(s, parent, opts)).join("");
+    return { result: wrapDidl(xml), numberReturned: sliced.length, totalMatches: total };
+  }
+
+  // ── Subsonic helpers ──────────────────────────────────────────────────
+
+  private async fetchArtist(rawId: string): Promise<SubsonicArtist | null> {
+    const body = await this.caller.call("/rest/getArtist", { id: `ar${rawId}` });
+    return (body["subsonic-response"].artist as SubsonicArtist | undefined) ?? null;
+  }
+
+  private async fetchAlbum(rawId: string): Promise<SubsonicAlbum | null> {
+    const body = await this.caller.call("/rest/getAlbum", { id: `al${rawId}` });
+    return (body["subsonic-response"].album as SubsonicAlbum | undefined) ?? null;
+  }
+
+  private songItem(s: SubsonicSong, parentId: string, opts: BrowseOptions): string {
+    const trackId = stripPrefix(s.id, "t");
+    return buildAudioItem({
+      objectId: trackId,
+      parentId,
+      trackId,
+      title: s.title,
+      artist: s.artist,
+      album: s.album,
+      albumArtUri: coverArtUri(opts.baseUrl, s.coverArt),
+      durationSec: s.duration ?? 0,
+      mimeType: mimeForContentType(s.contentType),
+      streamUri: streamUri(opts, trackId),
+      // DLNA.ORG_OP=01 indicates Range support; clients use it for seek.
+      protocolInfoExtras: DLNA_PROTOCOL_INFO_EXTRAS,
+    });
   }
 
   private staticChildren(items: string[], opts: BrowseOptions): BrowseResult {
@@ -338,229 +567,6 @@ export class DlnaObjectService {
       numberReturned: sliced.length,
       totalMatches: items.length,
     };
-  }
-
-  // ---------- DB queries ----------
-
-  private countArtists(): number {
-    // Hide artists with no release group of their own. They show up as
-    // empty containers in DLNA browsers (VLC, etc.) because the artist→
-    // album path filters by `urg.artist_id` and a track-only credit owns
-    // no release group. Their tracks remain reachable via the album they
-    // appear on.
-    const row = this.db
-      .prepare(
-        `SELECT COUNT(*) AS n FROM unified_artists ua
-          WHERE EXISTS (SELECT 1 FROM unified_release_groups urg
-                         WHERE urg.artist_id = ua.id)`,
-      )
-      .get() as { n: number };
-    return row.n;
-  }
-
-  private countAlbums(): number {
-    const row = this.db
-      .prepare("SELECT COUNT(*) AS n FROM unified_release_groups")
-      .get() as { n: number };
-    return row.n;
-  }
-
-  private countTracks(): number {
-    const row = this.db
-      .prepare("SELECT COUNT(*) AS n FROM unified_tracks")
-      .get() as { n: number };
-    return row.n;
-  }
-
-  private getArtist(id: string): ArtistRow | null {
-    return (
-      (this.db
-        .prepare(
-          `SELECT ua.id, ua.name, ua.image_url,
-                  (SELECT COUNT(*) FROM unified_release_groups urg WHERE urg.artist_id = ua.id) AS album_count
-             FROM unified_artists ua
-            WHERE ua.id = ?`,
-        )
-        .get(id) as ArtistRow | undefined) ?? null
-    );
-  }
-
-  private getAlbum(id: string): AlbumRow | null {
-    return (
-      (this.db
-        .prepare(
-          `SELECT urg.id, urg.name, urg.artist_id, ua.name AS artist_name,
-                  urg.image_url,
-                  (SELECT COUNT(*) FROM unified_tracks ut
-                     JOIN unified_releases ur ON ur.id = ut.release_id
-                    WHERE ur.release_group_id = urg.id) AS track_count
-             FROM unified_release_groups urg
-             JOIN unified_artists ua ON ua.id = urg.artist_id
-            WHERE urg.id = ?`,
-        )
-        .get(id) as AlbumRow | undefined) ?? null
-    );
-  }
-
-  private listArtists(opts: BrowseOptions): BrowseResult {
-    const total = this.countArtists();
-    const rows = this.db
-      .prepare(
-        `SELECT ua.id, ua.name, ua.image_url,
-                (SELECT COUNT(*) FROM unified_release_groups urg WHERE urg.artist_id = ua.id) AS album_count
-           FROM unified_artists ua
-          WHERE EXISTS (SELECT 1 FROM unified_release_groups urg
-                         WHERE urg.artist_id = ua.id)
-          ORDER BY ua.name COLLATE NOCASE
-          LIMIT ? OFFSET ?`,
-      )
-      .all(this.limit(opts), opts.startIndex) as ArtistRow[];
-
-    const xml = rows
-      .map((r) =>
-        buildContainer({
-          objectId: artistObjectId(r.id),
-          parentId: ARTISTS_ID,
-          title: r.name,
-          upnpClass: "object.container.person.musicArtist",
-          childCount: r.album_count,
-          albumArtUri: r.image_url,
-          artist: r.name,
-        }),
-      )
-      .join("");
-    return {
-      result: wrapDidl(xml),
-      numberReturned: rows.length,
-      totalMatches: total,
-    };
-  }
-
-  private listAlbums(artistId: string | null, opts: BrowseOptions): BrowseResult {
-    const where = artistId ? "WHERE urg.artist_id = ?" : "";
-    const totalRow = this.db
-      .prepare(
-        `SELECT COUNT(*) AS n FROM unified_release_groups urg ${where}`,
-      )
-      .get(...(artistId ? [artistId] : [])) as { n: number };
-    const total = totalRow.n;
-
-    const rows = this.db
-      .prepare(
-        `SELECT urg.id, urg.name, urg.artist_id, ua.name AS artist_name,
-                urg.image_url,
-                (SELECT COUNT(*) FROM unified_tracks ut
-                   JOIN unified_releases ur ON ur.id = ut.release_id
-                  WHERE ur.release_group_id = urg.id) AS track_count
-           FROM unified_release_groups urg
-           JOIN unified_artists ua ON ua.id = urg.artist_id
-           ${where}
-          ORDER BY urg.year IS NULL, urg.year, urg.name COLLATE NOCASE
-          LIMIT ? OFFSET ?`,
-      )
-      .all(...(artistId ? [artistId] : []), this.limit(opts), opts.startIndex) as AlbumRow[];
-
-    const xml = rows
-      .map((r) =>
-        buildContainer({
-          objectId: albumObjectId(r.id),
-          parentId: artistId ? artistObjectId(artistId) : ALBUMS_ID,
-          title: r.name,
-          upnpClass: "object.container.album.musicAlbum",
-          childCount: r.track_count,
-          albumArtUri: albumArtUriFor(opts.baseUrl, r.image_url),
-          artist: r.artist_name,
-        }),
-      )
-      .join("");
-    return {
-      result: wrapDidl(xml),
-      numberReturned: rows.length,
-      totalMatches: total,
-    };
-  }
-
-  private listTracks(
-    artistId: string | null,
-    albumId: string | null,
-    opts: BrowseOptions,
-  ): BrowseResult {
-    const filters: string[] = [];
-    const params: unknown[] = [];
-    if (artistId) {
-      filters.push("ut.artist_id = ?");
-      params.push(artistId);
-    }
-    if (albumId) {
-      filters.push("ur.release_group_id = ?");
-      params.push(albumId);
-    }
-    const where = filters.length ? `WHERE ${filters.join(" AND ")}` : "";
-
-    const totalRow = this.db
-      .prepare(
-        `SELECT COUNT(*) AS n
-           FROM unified_tracks ut
-           JOIN unified_releases ur ON ur.id = ut.release_id
-           ${where}`,
-      )
-      .get(...params) as { n: number };
-    const total = totalRow.n;
-
-    const rows = this.db
-      .prepare(
-        `SELECT ut.id, ut.title, ut.duration_ms,
-                ua.name AS artist_name,
-                urg.name AS album_name, urg.image_url AS album_art,
-                ts.format, ts.bitrate
-           FROM unified_tracks ut
-           JOIN unified_artists ua ON ua.id = ut.artist_id
-           JOIN unified_releases ur ON ur.id = ut.release_id
-           JOIN unified_release_groups urg ON urg.id = ur.release_group_id
-           LEFT JOIN track_sources ts ON ts.unified_track_id = ut.id AND ts.preferred = 1
-           ${where}
-          ORDER BY ut.disc_number, ut.track_number, ut.title COLLATE NOCASE
-          LIMIT ? OFFSET ?`,
-      )
-      .all(...params, this.limit(opts), opts.startIndex) as TrackRow[];
-
-    const parentId = albumId
-      ? albumObjectId(albumId)
-      : artistId
-        ? artistObjectId(artistId)
-        : TRACKS_ID;
-
-    const xml = rows
-      .map((r) =>
-        buildAudioItem({
-          objectId: r.id,
-          parentId,
-          trackId: r.id,
-          title: r.title,
-          artist: r.artist_name,
-          album: r.album_name,
-          albumArtUri: albumArtUriFor(opts.baseUrl, r.album_art),
-          durationSec: Math.floor((r.duration_ms ?? 0) / 1000),
-          mimeType: mimeForFormat(r.format),
-          streamUri: streamUri(opts, r.id),
-          // DLNA.ORG_OP=01 indicates Range support; clients use it for seek.
-          protocolInfoExtras: "DLNA.ORG_OP=01;DLNA.ORG_FLAGS=01700000000000000000000000000000",
-        }),
-      )
-      .join("");
-
-    return {
-      result: wrapDidl(xml),
-      numberReturned: rows.length,
-      totalMatches: total,
-    };
-  }
-
-  private limit(opts: BrowseOptions): number {
-    // UPnP convention: 0 = "as many as you can". Cap at 1000 so a buggy
-    // client can't ask for the whole table in one shot.
-    if (opts.requestedCount === 0) return 1000;
-    return Math.min(opts.requestedCount, 1000);
   }
 }
