+++ hub/test/dlna-objects.test.ts
@@ -1,7 +1,22 @@
-import { describe, it, expect, beforeEach } from "vitest";
-import Database from "better-sqlite3";
-import { readFileSync } from "node:fs";
-import { resolve } from "node:path";
+/**
+ * Unit/integration tests for the DLNA ContentDirectory service
+ * (`DlnaObjectService`), promoted from the #213 spike when the service was
+ * rewritten against the Subsonic HTTP client (#219).
+ *
+ * The service does no direct DB access — every browse / search call goes
+ * through a `SubsonicCaller`. The tests boot a real `buildApp()` and wire
+ * the service to an `app.inject()`-backed caller, which is also what
+ * production does (server.ts). Fixture data is seeded directly into the
+ * unified library tables; the Subsonic route handlers serialize it back
+ * out exactly as they do in production.
+ */
+import { describe, it, expect, beforeAll, afterAll } from "vitest";
+import { mkdtempSync, rmSync } from "node:fs";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+import type { FastifyInstance } from "fastify";
+import { buildApp } from "../src/server.js";
+import { setPassword } from "../src/auth/passwords.js";
 import {
   DlnaObjectService,
   parseObjectId,
@@ -11,44 +26,60 @@ import {
   MUSIC_ID,
   ARTISTS_ID,
   ALBUMS_ID,
-  TRACKS_ID,
+  type SubsonicCaller,
+  type SubsonicResponse,
 } from "../src/services/dlna-objects.js";
 
-const SCHEMA = readFileSync(
-  resolve(__dirname, "..", "src", "db", "schema.sql"),
-  "utf8",
-);
+const SUB_USER = "tester";
+const SUB_PASS = "testerpw";
+const BASE_URL = "http://lan:3000";
+const SECRET = Buffer.from("a".repeat(32));
 
-function seedDb(): Database.Database {
-  const db = new Database(":memory:");
-  db.exec(SCHEMA);
-  // One user (FK target for instances).
+const opts = {
+  startIndex: 0,
+  requestedCount: 0,
+  baseUrl: BASE_URL,
+  castSecret: SECRET,
+  username: "tester",
+};
+
+function seedFixtures(app: FastifyInstance): void {
+  const db = app.db;
+  const enc = setPassword(SUB_PASS, app.passwordKey);
   db.prepare(
-    "INSERT INTO users (id, username, is_admin) VALUES ('u', 'admin', 1)",
-  ).run();
-  // Synthetic local instance.
+    "INSERT INTO users (id, username, password_enc, is_admin) VALUES (?, ?, ?, 1)",
+  ).run("u-tester", SUB_USER, enc);
+
+  // buildApp() already seeds a `__system__` user + local instance row with
+  // musicfolder_id=1. Update that row so the fixtures below resolve against
+  // `instance_id='local'`.
   db.prepare(
-    `INSERT INTO instances (id, name, url, encrypted_credentials, owner_id, status)
-     VALUES ('local', 'local', 'http://nav', '', 'u', 'online')`,
+    `UPDATE instances
+       SET id = 'local', owner_id = 'u-tester', status = 'online'
+     WHERE musicfolder_id = 1`,
   ).run();
-  // Artist / album / release / track all wired together.
+
   db.prepare(
     `INSERT INTO unified_artists (id, name, name_normalized, image_url)
      VALUES ('a1', 'Artist One', 'artist one', NULL),
             ('a2', 'Artist Two', 'artist two', NULL),
             ('a3', 'Featured Only', 'featured only', NULL)`,
   ).run();
+
   db.prepare(
     `INSERT INTO unified_release_groups (id, name, name_normalized, artist_id, year)
      VALUES ('rg1', 'Album One', 'album one', 'a1', 2001),
             ('rg2', 'Album Two', 'album two', 'a1', 2002),
             ('rg3', 'Other Album', 'other album', 'a2', 2003)`,
   ).run();
+
   db.prepare(
-    `INSERT INTO unified_releases (id, release_group_id, name)
-     VALUES ('r1', 'rg1', 'Album One'),
-            ('r2', 'rg2', 'Album Two')`,
+    `INSERT INTO unified_releases (id, release_group_id, name, track_count)
+     VALUES ('r1', 'rg1', 'Album One', 2),
+            ('r2', 'rg2', 'Album Two', 1),
+            ('r3', 'rg3', 'Other Album', 0)`,
   ).run();
+
   // t3 credits 'a3' (the orphan) on 'a1'-owned rg2, simulating a featured-
   // artist credit. 'a3' should NOT appear in the artist list.
   db.prepare(
@@ -57,7 +88,7 @@ function seedDb(): Database.Database {
             ('t2', 'Track Two', 'track two', 'r1', 'a1', 2, 200000),
             ('t3', 'Guest Spot', 'guest spot', 'r2', 'a3', 1, 150000)`,
   ).run();
-  // Make t1 streamable via a preferred local source.
+
   db.prepare(
     `INSERT INTO instance_artists (id, instance_id, remote_id, name)
      VALUES ('local:ar1', 'local', 'ar1', 'Artist One')`,
@@ -76,7 +107,30 @@ function seedDb(): Database.Database {
      VALUES ('ts1', 't1', 'local', 'local:tr1', 'mp3', 256, 1),
             ('ts2', 't2', 'local', 'local:tr2', 'flac', 1000, 1)`,
   ).run();
-  return db;
+}
+
+/** Subsonic caller that drives the hub's own `/rest/*` via `app.inject()`. */
+function makeInjectCaller(app: FastifyInstance): SubsonicCaller {
+  return {
+    async call(endpoint, params): Promise<SubsonicResponse> {
+      const qs = new URLSearchParams({
+        u: SUB_USER,
+        p: SUB_PASS,
+        f: "json",
+        v: "1.16.1",
+        c: "dlna-test",
+        ...params,
+      });
+      const res = await app.inject({
+        method: "GET",
+        url: `${endpoint}?${qs.toString()}`,
+      });
+      if (res.statusCode !== 200) {
+        throw new Error(`${endpoint} → ${res.statusCode}`);
+      }
+      return res.json() as SubsonicResponse;
+    },
+  };
 }
 
 describe("parseObjectId", () => {
@@ -85,86 +139,111 @@ describe("parseObjectId", () => {
     expect(parseObjectId(MUSIC_ID)?.kind).toBe("music");
     expect(parseObjectId(ARTISTS_ID)?.kind).toBe("artists");
     expect(parseObjectId(ALBUMS_ID)?.kind).toBe("albums");
-    expect(parseObjectId(TRACKS_ID)?.kind).toBe("tracks");
   });
   it("parses dynamic artist/album IDs", () => {
-    expect(parseObjectId(artistObjectId("a1"))).toEqual({
-      kind: "artist",
-      id: "a1",
-    });
-    expect(parseObjectId(albumObjectId("rg1"))).toEqual({
-      kind: "album",
-      id: "rg1",
-    });
+    expect(parseObjectId(artistObjectId("a1"))).toEqual({ kind: "artist", id: "a1" });
+    expect(parseObjectId(albumObjectId("rg1"))).toEqual({ kind: "album", id: "rg1" });
   });
   it("returns null for unknown IDs", () => {
     expect(parseObjectId("nope")).toBeNull();
   });
+  it("returns null for the retired All Tracks container ID", () => {
+    // #219: "All Tracks" was dropped from the hierarchy — Subsonic has no
+    // global track enumeration endpoint and the only synthesis is N+1.
+    expect(parseObjectId("0/music/tracks")).toBeNull();
+  });
 });
 
-describe("DlnaObjectService.browse", () => {
-  let db: Database.Database;
+describe("DlnaObjectService.browse (Subsonic-backed)", () => {
+  let app: FastifyInstance;
   let svc: DlnaObjectService;
-  // #218: DIDL `res@uri` is now a self-contained cast-token URL.
-  // Tests below assert the prefix only — the cast token body changes per run.
-  const opts = {
-    startIndex: 0,
-    requestedCount: 0,
-    baseUrl: "http://lan:3000",
-    castSecret: Buffer.from("a".repeat(32)),
-    username: "tester",
-  };
+  const tmp = mkdtempSync(join(tmpdir(), "poutine-dlna-svc-"));
 
-  beforeEach(() => {
-    db = seedDb();
-    svc = new DlnaObjectService(db);
+  beforeAll(async () => {
+    app = await buildApp({
+      databasePath: ":memory:",
+      jwtSecret: "x",
+      poutinePrivateKeyPath: join(tmp, "ed.pem"),
+      poutinePasswordKeyPath: join(tmp, "pwkey"),
+      poutineInstanceId: "dlna-svc-test",
+      poutineOwnerUsername: "owner-unused",
+    });
+    await app.ready();
+    seedFixtures(app);
+    svc = new DlnaObjectService(makeInjectCaller(app));
   });
 
-  it("root → BrowseDirectChildren lists the Music container", () => {
-    const out = svc.browse(ROOT_ID, "BrowseDirectChildren", opts);
+  afterAll(async () => {
+    await app.close();
+    rmSync(tmp, { recursive: true, force: true });
+  });
+
+  // ── BrowseDirectChildren ──────────────────────────────────────────────
+
+  it("root → BrowseDirectChildren lists the Music container", async () => {
+    const out = await svc.browse(ROOT_ID, "BrowseDirectChildren", opts);
     expect(out.numberReturned).toBe(1);
     expect(out.totalMatches).toBe(1);
     expect(out.result).toContain(`id="${MUSIC_ID}"`);
     expect(out.result).toContain("<dc:title>Music</dc:title>");
   });
 
-  it("artists → returns one container per unified artist, ordered by name", () => {
-    const out = svc.browse(ARTISTS_ID, "BrowseDirectChildren", opts);
-    expect(out.totalMatches).toBe(2);
+  it("music → BrowseDirectChildren lists Artists + Albums (no All Tracks)", async () => {
+    const out = await svc.browse(MUSIC_ID, "BrowseDirectChildren", opts);
     expect(out.numberReturned).toBe(2);
+    expect(out.result).toContain(`id="${ARTISTS_ID}"`);
+    expect(out.result).toContain(`id="${ALBUMS_ID}"`);
+    // #219: All Tracks container is intentionally absent.
+    expect(out.result).not.toContain("All Tracks");
+    expect(out.result).not.toContain("0/music/tracks");
+  });
+
+  it("artists → returns one container per unified artist, ordered by name", async () => {
+    const out = await svc.browse(ARTISTS_ID, "BrowseDirectChildren", opts);
+    expect(out.numberReturned).toBe(2);
+    expect(out.totalMatches).toBe(2);
     const idx1 = out.result.indexOf("Artist One");
     const idx2 = out.result.indexOf("Artist Two");
     expect(idx1).toBeGreaterThan(-1);
     expect(idx2).toBeGreaterThan(idx1);
     expect(out.result).toContain(`id="${artistObjectId("a1")}"`);
+    expect(out.result).toContain(`id="${artistObjectId("a2")}"`);
   });
 
-  it("artists → excludes track-only credits with no release group of their own", () => {
-    const out = svc.browse(ARTISTS_ID, "BrowseDirectChildren", opts);
-    expect(out.totalMatches).toBe(2);
+  it("artists → excludes track-only credits with no release group of their own", async () => {
+    const out = await svc.browse(ARTISTS_ID, "BrowseDirectChildren", opts);
     expect(out.result).not.toContain("Featured Only");
   });
 
-  it("artist/<id> → lists release groups for that artist", () => {
-    const out = svc.browse(artistObjectId("a1"), "BrowseDirectChildren", opts);
+  it("albums → BrowseDirectChildren returns release groups via getAlbumList2", async () => {
+    const out = await svc.browse(ALBUMS_ID, "BrowseDirectChildren", opts);
+    expect(out.numberReturned).toBe(3);
+    expect(out.result).toContain(`id="${albumObjectId("rg1")}"`);
+    expect(out.result).toContain(`id="${albumObjectId("rg2")}"`);
+    expect(out.result).toContain(`id="${albumObjectId("rg3")}"`);
+    // Subsonic getAlbumList2 has no total; we report -1.
+    expect(out.totalMatches).toBe(-1);
+  });
+
+  it("artist/<id> → lists release groups for that artist", async () => {
+    const out = await svc.browse(artistObjectId("a1"), "BrowseDirectChildren", opts);
     expect(out.totalMatches).toBe(2);
     expect(out.result).toContain("Album One");
     expect(out.result).toContain("Album Two");
     expect(out.result).toContain(`id="${albumObjectId("rg1")}"`);
   });
 
-  it("album/<id> → lists tracks for that release group, ordered by disc/track", () => {
-    const out = svc.browse(albumObjectId("rg1"), "BrowseDirectChildren", opts);
+  it("album/<id> → lists tracks in disc/track order with stream URLs + MIME", async () => {
+    const out = await svc.browse(albumObjectId("rg1"), "BrowseDirectChildren", opts);
     expect(out.totalMatches).toBe(2);
     const idx1 = out.result.indexOf("Track One");
     const idx2 = out.result.indexOf("Track Two");
     expect(idx1).toBeGreaterThan(-1);
     expect(idx2).toBeGreaterThan(idx1);
-    // #218: stream URL now points at Hub's Subsonic /rest/stream.view
-    // with id=t<unifiedId>, an embedded castToken, and dlna=1 marker.
-    expect(out.result).toContain("http://lan:3000/rest/stream.view?id=tt1");
-    expect(out.result).toContain("http://lan:3000/rest/stream.view?id=tt2");
-    // `&` is XML-escaped to `&amp;` in DIDL — match raw token names.
+    // #218: stream URL points at Hub's Subsonic /rest/stream.view with
+    // id=t<unifiedId>, an embedded castToken, and dlna=1 marker.
+    expect(out.result).toContain(`${BASE_URL}/rest/stream.view?id=tt1`);
+    expect(out.result).toContain(`${BASE_URL}/rest/stream.view?id=tt2`);
     expect(out.result).toContain("castToken=");
     expect(out.result).toContain("dlna=1");
     expect(out.result).toContain("&amp;");
@@ -173,17 +252,50 @@ describe("DlnaObjectService.browse", () => {
     );
     // FLAC source → audio/flac MIME.
     expect(out.result).toContain("audio/flac");
+    expect(out.result).toContain("audio/mpeg");
   });
 
-  it("BrowseMetadata on root returns a single container describing root", () => {
-    const out = svc.browse(ROOT_ID, "BrowseMetadata", opts);
+  // ── BrowseMetadata ────────────────────────────────────────────────────
+
+  it("BrowseMetadata on root returns a single container describing root", async () => {
+    const out = await svc.browse(ROOT_ID, "BrowseMetadata", opts);
     expect(out.numberReturned).toBe(1);
     expect(out.result).toContain(`id="${ROOT_ID}"`);
     expect(out.result).toContain('parentID="-1"');
   });
 
-  it("pagination via startIndex/requestedCount", () => {
-    const page1 = svc.browse(ARTISTS_ID, "BrowseDirectChildren", {
+  it("BrowseMetadata(artist/a1) returns one container with album count", async () => {
+    const out = await svc.browse(artistObjectId("a1"), "BrowseMetadata", opts);
+    expect(out.numberReturned).toBe(1);
+    expect(out.result).toContain(`id="${artistObjectId("a1")}"`);
+    expect(out.result).toContain("Artist One");
+  });
+
+  it("BrowseMetadata(album/rg1) returns one container with song count", async () => {
+    const out = await svc.browse(albumObjectId("rg1"), "BrowseMetadata", opts);
+    expect(out.numberReturned).toBe(1);
+    expect(out.result).toContain(`id="${albumObjectId("rg1")}"`);
+    expect(out.result).toContain("Album One");
+  });
+
+  // ── Search ───────────────────────────────────────────────────────────
+
+  it("Search by title surfaces a song item with a stream URL", async () => {
+    const out = await svc.search("Track One", { ...opts, requestedCount: 10 });
+    expect(out.numberReturned).toBeGreaterThan(0);
+    expect(out.result).toContain("Track One");
+    expect(out.result).toContain(`${BASE_URL}/rest/stream.view?id=tt1`);
+  });
+
+  it("Search by artist surfaces an artist container", async () => {
+    const out = await svc.search("Artist One", { ...opts, requestedCount: 10 });
+    expect(out.result).toContain("Artist One");
+  });
+
+  // ── Pagination ───────────────────────────────────────────────────────
+
+  it("pagination via startIndex/requestedCount on artists", async () => {
+    const page1 = await svc.browse(ARTISTS_ID, "BrowseDirectChildren", {
       ...opts,
       requestedCount: 1,
     });
@@ -192,7 +304,7 @@ describe("DlnaObjectService.browse", () => {
     expect(page1.result).toContain("Artist One");
     expect(page1.result).not.toContain("Artist Two");
 
-    const page2 = svc.browse(ARTISTS_ID, "BrowseDirectChildren", {
+    const page2 = await svc.browse(ARTISTS_ID, "BrowseDirectChildren", {
       ...opts,
       startIndex: 1,
       requestedCount: 1,
@@ -201,10 +313,22 @@ describe("DlnaObjectService.browse", () => {
     expect(page2.result).toContain("Artist Two");
   });
 
-  it("unknown object ID returns an empty DIDL envelope", () => {
-    const out = svc.browse("nope", "BrowseDirectChildren", opts);
+  // ── Error cases ──────────────────────────────────────────────────────
+
+  it("unknown object ID returns an empty DIDL envelope", async () => {
+    const out = await svc.browse("nope", "BrowseDirectChildren", opts);
     expect(out.numberReturned).toBe(0);
     expect(out.totalMatches).toBe(0);
     expect(out.result).toMatch(/<DIDL-Lite[^/]*\/>|<DIDL-Lite[^>]*><\/DIDL-Lite>/);
+  });
+
+  it("missing artist returns an empty DIDL envelope", async () => {
+    const out = await svc.browse(artistObjectId("nosuch"), "BrowseDirectChildren", opts);
+    expect(out.numberReturned).toBe(0);
+  });
+
+  it("missing album returns an empty DIDL envelope", async () => {
+    const out = await svc.browse(albumObjectId("nosuch"), "BrowseDirectChildren", opts);
+    expect(out.numberReturned).toBe(0);
   });
 });
